# dashboard/swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "A space station except way more Aero-dynamic",
    "title": "Aerostation API service",
    "contact": {
      "email": "sbhagat@aerospike.com"
    },
    "version": "1.0.0"
  },
  "basePath": "/v1",
  "paths": {
    "/api/v1/admin/aerospike/clusters": {
      "get": {
        "description": "Get names of all the workload clusters in array of string",
        "produces": [
          "application/json"
        ],
        "tags": [
          "admin"
        ],
        "summary": "Get all aerospike database clusters",
        "operationId": "getAllAerospikeClusters",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/NamespacedName"
              },
              "example": [
                {
                  "name": "cluster-name",
                  "namespace": "cluster-namespace"
                },
                {
                  "name": "cluster2-name",
                  "namespace": "cluster2-namespace"
                }
              ]
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Create a aerospike database cluster",
        "produces": [
          "application/json"
        ],
        "tags": [
          "admin"
        ],
        "summary": "Create a aerospike database cluster",
        "operationId": "adminCreateAerospikeCluster",
        "parameters": [
          {
            "description": "Request body for creating workload cluster",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAerospikeClusterAdminRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aerospike database created successfully"
          },
          "400": {
            "description": "Unable to create cluster"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Unable to create cluster"
          }
        }
      }
    },
...
    "/api/v1/aerospike/clusters": {
      "get": {
        "description": "Get names of all the workload clusters in array of string",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Get all aerospike database clusters",
        "operationId": "getAerospikeClusters",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "example": [
                "aerospike-cluster-1",
                "aerospike-cluster-2"
              ]
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Create a aerospike database cluster",
        "produces": [
          "application/json"
        ],
        "tags": [
          "aerospike"
        ],
        "summary": "Create a aerospike database cluster",
        "operationId": "createAerospikeCluster",
        "parameters": [
          {
            "description": "Request body for creating workload cluster",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAerospikeClusterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aerospike database created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Unable to create cluster"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateAerospikeClusterAdminRequest": {
      "description": "CreateAerospikeClusterAdminRequest A request to create an aerospike cluster",
      "type": "object",
      "required": [
        "kubernetesClusterName",
        "kubernetesClusterNamespace",
        "name",
        "namespace",
        "databaseType",
        "options"
      ],
      "properties": {
        "databaseType": {
          "description": "DatabaseType of the aerospike cluster (memory/ssd/performance)",
          "type": "string",
          "x-go-name": "DatabaseType",
          "example": "memory"
        },
        "kubernetesClusterName": {
          "description": "KubernetesClusterName of workload cluster",
          "type": "string",
          "x-go-name": "KubernetesClusterName",
          "example": "cluster-1"
        },
        "kubernetesClusterNamespace": {
          "description": "KubernetesClusterNamespace of workload cluster",
          "type": "string",
          "x-go-name": "KubernetesClusterNamespace",
          "example": "default"
        },
        "name": {
          "description": "Name of aerospike cluster",
          "type": "string",
          "x-go-name": "Name",
          "example": "name"
        },
        "namespace": {
          "description": "Namespace of aerospike cluster",
          "type": "string",
          "x-go-name": "Namespace",
          "example": "default"
        },
        "options": {
          "$ref": "#/definitions/DatabaseOptions"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api-server/pkg/requests"
    },
    "CreateAerospikeClusterRequest": {
      "description": "CreateAerospikeClusterRequest A request to create an aerospike cluster",
      "type": "object",
      "required": [
        "region",
        "name",
        "databaseType",
        "options"
      ],
      "properties": {
        "databaseType": {
          "description": "DatabaseType of the aerospike cluster (memory/ssd/performance)",
          "type": "string",
          "x-go-name": "DatabaseType",
          "example": "memory"
        },
        "name": {
          "description": "Name of aerospike cluster",
          "type": "string",
          "x-go-name": "Name",
          "example": "name"
        },
        "options": {
          "$ref": "#/definitions/DatabaseOptions"
        },
        "region": {
          "description": "KubernetesClusterName of workload cluster",
          "type": "string",
          "x-go-name": "Region",
          "example": "cluster-1"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api-server/pkg/requests"
    },
    "DatabaseOptions": {
      "description": "DatabaseOptions the mutable options of an aerospike cluster",
      "type": "object",
      "required": [
        "replicas"
      ],
      "properties": {
        "replicas": {
          "description": "Replicas of the aerospike cluster",
          "type": "integer",
          "format": "int32",
          "x-go-name": "Replicas",
          "example": 2
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api/v1"
    },
...
    "NamespacedName": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace",
          "example": "default"
        }
      },
      "x-go-package": "github.com/aerospike/aerostation/api/v1"
    },
...
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "BasicAuth": []
    }
  ]
}